+++ package.json
@@ -2,18 +2,24 @@
   "name": "tributejs",
   "version": "1.1.2",
   "description": "Native ES6 @mentions",
+  "main" : "dist/tribute.js"
   "devDependencies": {
     "babel-cli": "~6.0.0",
     "babel-preset-es2015": "~6.3.13",
+    "babelify": "^7.3.0",
+    "browserify": "^13.0.1",
     "gulp-babel": "^6.1.2",
     "gulp-plumber": "^1.1.0",
     "gulp-sass": "^2.2.0",
+    "gulp-uglifyjs": "^0.6.2",
     "gulp-webserver": "^0.9.1",
     "jasmine-core": "^2.4.1",
     "karma": "^0.13.19",
     "karma-chrome-launcher": "^0.2.2",
     "karma-jasmine": "^0.3.7",
     "opn": "^4.0.0",
+    "vinyl-source-stream": "^1.1.0",
+    "watchify": "^3.7.0",
     "zombie": "^4.2.1"
   },
   "repository": {
