+++ src/Tribute.js
@@ -1,0 +1,218 @@
+import "./utils";
+import TributeEvents from "./TributeEvents";
+import TributeMenuEvents from "./TributeMenuEvents";
+import TributeRange from "./TributeRange";
+import TributeSearch from "./TributeSearch";
+
+class Tribute {
+    constructor({
+        values = null,
+        iframe = null,
+        selectClass = 'highlight',
+        trigger = '@',
+        selectTemplate = null,
+        menuItemTemplate = null,
+        lookup = 'key',
+        fillAttr = 'value',
+        collection = null,
+        menuContainer = null
+    }) {
+
+        this.menuSelected = 0
+        this.current = {}
+        this.inputEvent = false
+        this.isActive = false
+        this.menuContainer = menuContainer
+
+        if (values) {
+            this.collection = [{
+                // symbol that starts the lookup
+                trigger: trigger,
+
+                iframe: iframe,
+
+                selectClass: selectClass,
+
+                // function called on select that retuns the content to insert
+                selectTemplate: (selectTemplate || Tribute.defaultSelectTemplate).bind(this),
+
+                menuItemTemplate: (menuItemTemplate || Tribute.defaultMenuItemTemplate).bind(this),
+
+                // column to search against in the object
+                lookup: lookup,
+
+                // column that contains the content to insert by default
+                fillAttr: fillAttr,
+
+                // array of objects
+                values: values
+            }]
+        }
+        else if (collection) {
+            this.collection = collection.map(item => {
+                return {
+                    trigger: item.trigger || trigger,
+                    iframe: item.iframe || iframe,
+                    selectClass: item.selectClass || selectClass,
+                    selectTemplate: (item.selectTemplate || Tribute.defaultSelectTemplate).bind(this),
+                    menuItemTemplate: (item.menuItemTemplate || Tribute.defaultMenuItemTemplate).bind(this),
+                    lookup: item.lookup || lookup,
+                    fillAttr: item.fillAttr || fillAttr,
+                    values: item.values
+                }
+            })
+        }
+        else {
+            throw new Error('[Tribute] No collection specified.')
+        }
+
+        new TributeRange(this);
+        new TributeEvents(this);
+        new TributeMenuEvents(this);
+        new TributeSearch(this);
+    }
+
+    static defaultSelectTemplate(item) {
+        return `@${item.original[this.current.collection.fillAttr]}`
+    }
+
+    static defaultMenuItemTemplate(matchItem) {
+        return matchItem.string
+    }
+
+    static inputTypes() {
+        return ['TEXTAREA', 'INPUT']
+    }
+
+    triggers() {
+        return this.collection.map(config => {
+            return config.trigger
+        })
+    }
+
+    attach(el) {
+        if (!el) {
+            throw new Error('[Tribute] Must pass in a DOM node or NodeList.')
+        }
+
+        // Check if it is a jQuery collection
+        if (typeof jQuery !== 'undefined' && el instanceof jQuery) {
+            el = el.get()
+        }
+
+        // Is el an Array/Array-like object?
+        if (el.constructor === NodeList || el.constructor === HTMLCollection || el.constructor === Array) {
+            let length = el.length
+            for (var i = 0; i < length; ++i) {
+                this._attach(el[i])
+            }
+        }
+        else {
+            this._attach(el)
+        }
+    }
+
+    _attach(el) {
+        if (el.hasAttribute('data-tribute')) {
+            console.warn('Tribute was already bound to ' + el.nodeName)
+        }
+
+        this.ensureEditable(el)
+        this.events.bind(el)
+        el.setAttribute('data-tribute', true)
+    }
+
+    ensureEditable(element) {
+        if (Tribute.inputTypes().indexOf(element.nodeName) === -1) {
+            if (element.contentEditable) {
+                element.contentEditable = true
+            }
+            else {
+                throw new Error('[Tribute] Cannot bind to ' + element.nodeName)
+            }
+        }
+    }
+
+    createMenu() {
+        let wrapper = this.range.getDocument().createElement('div'),
+            ul = this.range.getDocument().createElement('ul')
+
+        wrapper.className = 'tribute-container'
+        wrapper.appendChild(ul)
+
+        if (this.menuContainer) {
+            return this.menuContainer.appendChild(wrapper)
+        }
+
+        return this.range.getDocument().body.appendChild(wrapper)
+    }
+
+    showMenuFor(element, scrollTo) {
+        let items
+            // create the menu if it doesn't exist.
+        if (!this.menu) {
+            this.menu = this.createMenu()
+            this.menuEvents.bind(this.menu)
+        }
+
+        this.isActive = true
+        this.menuSelected = 0
+
+        if (!this.current.mentionText) {
+            this.current.mentionText = ''
+        }
+
+        items = this.search.filter(this.current.mentionText, this.current.collection.values, {
+            pre: '<span>',
+            post: '</span>',
+            extract: (el) => {
+                return el[this.current.collection.lookup]
+            }
+        })
+
+        this.current.filteredItems = items
+
+        if (!items.length) {
+            this.hideMenu()
+            return
+        }
+
+        let ul = this.menu.querySelector('ul')
+
+        ul.innerHTML = ''
+
+        items.forEach((item, index) => {
+            let li = this.range.getDocument().createElement('li')
+            li.setAttribute('data-index', index)
+            if (this.menuSelected === index) {
+                li.className = this.current.collection.selectClass
+            }
+            li.innerHTML = this.current.collection.menuItemTemplate(item)
+            ul.appendChild(li)
+        })
+
+        this.range.positionMenuAtCaret(scrollTo)
+
+    }
+
+    hideMenu() {
+        if (this.menu) {
+            this.menu.style.cssText = 'display: none;'
+            this.isActive = false
+            this.menuSelected = 0
+            this.current = {}
+        }
+    }
+
+    selectItemAtIndex(index) {
+        let item = this.current.filteredItems[index]
+        let content = this.current.collection.selectTemplate(item)
+        this.replaceText(content)
+    }
+
+    replaceText(content) {
+        this.range.replaceTriggerText(content, true, true)
+    }
+}
+
+global.Tribute = Tribute;
